plugin-space: extract the shared related-object card
The Related companion and the record article's inline section carried
byte-identical copies of the masonry tile. Extract it as `RelatedObjectCard`.

TypeArticle's tile stays its own component: it is selectable, gates its body on
CardAnnotation, and takes its menu from callbacks rather than the object graph.
