+++ packages/plugins/plugin-space/src/components/RelatedObjectCard/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './RelatedObjectCard';
