+++ packages/plugins/plugin-space/src/components/index.ts
@@ -5,6 +5,7 @@
 export * from './AwaitingObject';
 export * from './CreateObjectPanel';
 export * from './ForeignKeys';
+export * from './RelatedObjectCard';
 export * from './RelatedTypeFilter';
 export * from './SyncStatus';
