+++ packages/plugins/plugin-space/src/containers/RelatedArticle/RelatedArticle.tsx
@@ -2,18 +2,15 @@
 // Copyright 2025 DXOS.org
 //
 
-import React, { useMemo } from 'react';
+import React from 'react';
 
-import { Surface } from '@dxos/app-framework/ui';
-import { AppSurface, useCardPivot, useObjectMenuItems } from '@dxos/app-toolkit/ui';
-import { Entity, Obj } from '@dxos/echo';
-import { Card, Icon, IconButton, Panel, Toolbar, useTranslation } from '@dxos/react-ui';
+import { type AppSurface } from '@dxos/app-toolkit/ui';
+import { Obj } from '@dxos/echo';
+import { Panel, Toolbar } from '@dxos/react-ui';
 import { Masonry } from '@dxos/react-ui-masonry';
-import { Menu } from '@dxos/react-ui-menu';
 
-import { RelatedTypeFilter } from '#components';
+import { RelatedObjectCard, RelatedTypeFilter } from '#components';
 import { useRelatedObjects, useRelatedTypeFilter } from '#hooks';
-import { meta } from '#meta';
 
 export type RelatedArticleProps = Pick<
   AppSurface.ObjectArticleProps<Obj.Unknown, {}, Obj.Unknown>,
@@ -28,7 +25,7 @@ export const RelatedArticle = ({ role, companionTo }: RelatedArticleProps) => {
   const { types, items, toggle } = useRelatedTypeFilter(related, contextId);
 
   return (
-    <Masonry.Root Tile={ObjectCard}>
+    <Masonry.Root Tile={RelatedObjectCard}>
       <Panel.Root role={role}>
         <Panel.Toolbar asChild>
           <Toolbar.Root>
@@ -42,44 +39,6 @@ export const RelatedArticle = ({ role, companionTo }: RelatedArticleProps) => {
         </Panel.Content>
       </Panel.Root>
     </Masonry.Root>
-  );
-};
-
-/** Masonry tile renderer for a related entity. */
-const ObjectCard = ({ data: subject, classNames }: { data: Entity.Unknown; classNames?: string }) => {
-  const { t } = useTranslation(meta.profile.key);
-  const data = useMemo(() => ({ subject }), [subject]);
-  const icon = Entity.getIcon(subject)?.icon ?? 'ph--circle-dashed--regular';
-
-  // The card menu renders in a portal; resolve the origin plank from the card element instead.
-  const [cardRef, pivotId] = useCardPivot();
-  const menuItems = useObjectMenuItems(subject, pivotId);
-
-  return (
-    <Menu.Root>
-      <Card.Root ref={cardRef} classNames={classNames}>
-        <Card.Header>
-          <Card.Block>
-            <Icon icon={icon} />
-          </Card.Block>
-          <Card.Title>{Entity.getLabel(subject, { fallback: 'typename' })}</Card.Title>
-          <Card.Block end>
-            <Menu.Trigger asChild disabled={!menuItems?.length}>
-              <IconButton
-                iconOnly
-                variant='ghost'
-                icon='ph--dots-three-vertical--regular'
-                label={t('more-actions.label')}
-              />
-            </Menu.Trigger>
-            <Menu.Content items={menuItems} />
-          </Card.Block>
-        </Card.Header>
-        <Card.Body>
-          <Surface.Surface type={AppSurface.CardContent} data={data} limit={1} />
-        </Card.Body>
-      </Card.Root>
-    </Menu.Root>
   );
 };
